# perf: parallel search partitions in subindex