V1 baseline
```
Running query with nprobes: 1, refine_factor: 0
Time: 652.877µs

Running query with nprobes: 10, refine_factor: 0
Time: 818.418µs

Running query with nprobes: 50, refine_factor: 0
Time: 1.406433ms

Running query with nprobes: 100, refine_factor: 0
Time: 1.747243ms

Running query with nprobes: 512, refine_factor: 0
Time: 5.754752ms
```

V3 after fix
```
Running query with nprobes: 1, refine_factor: 0
Time: 667.621µs

Running query with nprobes: 10, refine_factor: 0
Time: 707.589µs

Running query with nprobes: 50, refine_factor: 0
Time: 1.213155ms

Running query with nprobes: 100, refine_factor: 0
Time: 1.559892ms

Running query with nprobes: 512, refine_factor: 0
Time: 5.597119ms
```